+++ public/javascripts/jquery.wall.js
@@ -25,8 +25,8 @@
         className: "hint"
     };
 
-    $.fn.hintReplys = function(options) {
-        var opts = $.extend({}, $.fn.hintReplys.defaults, options);
+    $.fn.hintReplies = function(options) {
+        var opts = $.extend({}, $.fn.hintReplies.defaults, options);
 
         return this.map(function(){
                     jQuery(this).data("_default", jQuery(this).val());
@@ -63,7 +63,7 @@
                 .addClass('hint');
     };
 
-    $.fn.hintReplys.defaults = {
+    $.fn.hintReplies.defaults = {
         className: "hint",
         message: "Escribe tu commentario...",
         rows: 1
@@ -134,7 +134,7 @@
 
 if (window.Widget == undefined) window.Widget = {};
 
-// For collapsing some replys to display later if the user wants
+// For collapsing some replies to display later if the user wants
 Widget.Collapsable = Class.create({
     initialize: function(comment, collapsed, options)
     {
@@ -145,8 +145,8 @@ Widget.Collapsable = Class.create({
             'children_showed' : 2 //2 children showed always
         }).update(options);
 
-        this.replys = this.comment.select("div[id^=reply]");
-        this.total_replys = this.replys.size();
+        this.replies = this.comment.select("div[id^=reply]");
+        this.total_replies = this.replies.size();
         if (this.collapsed)
             this.collapsed.widget = this;
         this.init_state();
@@ -154,7 +154,7 @@ Widget.Collapsable = Class.create({
 
     init_state: function()
     {
-        if (this.total_replys > this.options.get('collapse_since')){
+        if (this.total_replies > this.options.get('collapse_since')){
           this.collapse();
           this.collapsed.show();
         }
@@ -164,8 +164,8 @@ Widget.Collapsable = Class.create({
     collapse: function()
     {
         // Se muestran solo los últimos children_showed, los primeros se ocultan
-        var hide_first = this.total_replys - this.options.get('children_showed');
-        this.replys.each(function(el,i){
+        var hide_first = this.total_replies - this.options.get('children_showed');
+        this.replies.each(function(el,i){
             if (i < hide_first) { el.hide(); };
         });
         return true;
@@ -173,8 +173,8 @@ Widget.Collapsable = Class.create({
 
     uncollapse: function()
     {
-        var hide_first = this.total_replys - this.options.get('children_showed');
-        this.replys.each(function(el,i){
+        var hide_first = this.total_replies - this.options.get('children_showed');
+        this.replies.each(function(el,i){
             if (i < hide_first) { el.show(); };
         });
         this.collapsed.hide(); // we can't use this.collapsed.addClassName("hide"); in IE6
@@ -182,7 +182,7 @@ Widget.Collapsable = Class.create({
     }
 });
 
-CollapsableReplys = Behavior.create({
+CollapsableReplies = Behavior.create({
     initialize : function(options) {
         this.options = $H({
             'selector_for_collapsed'  : "div[id^=collapsed]",